refactor: [Task 1.2] Extract team abbreviations to components module
