+++ src/lib.rs
@@ -54,9 +54,8 @@ pub use config::Config;
 pub use data_fetcher::api::fetch_liiga_data;
 pub use data_fetcher::models::{DetailedGameResponse, GameData, ScheduleResponse};
 pub use error::AppError;
-pub use teletext_ui::{
-    CompactDisplayConfig, GameResultData, TeletextPage, TeletextPageConfig, get_team_abbreviation,
-};
+pub use teletext_ui::{CompactDisplayConfig, GameResultData, TeletextPage, TeletextPageConfig};
+pub use ui::components::abbreviations::get_team_abbreviation;
 
 // Re-export cache debugging and monitoring functions for external tools
 pub use data_fetcher::cache::{
