+++ src/ui/mod.rs
@@ -1,3 +1,4 @@
+pub mod components;
 pub mod interactive;
 pub mod teletext;
